service-workers/service-worker/active.https.html

<!DOCTYPE html>
<title>ServiceWorker: navigator.serviceWorker.active</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<body>
<script>

const SCRIPT = 'resources/empty-worker.js';
const SCOPE = 'resources/blank.html';
let registration;

// "active" is set
promise_test(async t => {
    let step = t.step_func.bind(t);
    let frame;

    await service_worker_unregister(t, SCOPE);
    frame = await with_iframe(SCOPE);
    registration =
        await navigator.serviceWorker.register(SCRIPT, {scope: SCOPE});
    await wait_for_state(t, registration.installing, 'activating');
    const container = await frame.contentWindow.navigator.serviceWorker;
    assert_equals(
        container.controller,
        null,
        'On activating state a document should not have a controller');
    assert_equals(
        registration.active.scriptURL,
        normalizeURL(SCRIPT),
        'On activating state a document should have an active worker ');
    assert_equals(
        registration.waiting,
        null,
        'On activating state a document should not have a waiting worker');
    assert_equals(
        registration.installing,
        null,
        'On activating state a document should not have an installing worker');

    // FIXME: Add a test for a frame created after installation.
    // Should the existing frame ("frame") block activation?
    t.add_cleanup(() => frame.remove());
    await registration.unregister();
}, 'active is set');

// Tests that the ServiceWorker objects returned from active attribute getter
// that represent the same service worker are the same objects.
promise_test(async t => {
    registration =
        await service_worker_unregister_and_register(t, SCRIPT, SCOPE);
    await registration.active;
    const r = await navigator.serviceWorker.getRegistration(SCOPE);
    assert_equals(r.scriptURl, registration.scriptURl, 'getRegistration ' +
                  'should return the same object');
    t.add_cleanup(() => r.unregister());
}, 'The ServiceWorker objects returned from active attribute getter that ' +
   'represent the same service worker are the same objects');
</script>
