third_party/WebKit/LayoutTests/platform/win/external/wpt/service-workers/service-worker/active.https-expected.txt

This is a testharness.js-based test.
PASS active is set 
Harness: the test ran to completion.
